Rewire copyLocationProperty
